Create Brain Teaser2
In a session, for a school pattern such as Aptech, we have structured and unstructure codes. The styler follows a certain html standard. At a certain paint in time in the header figure was miss-interpreted by the  styler who failed to follow the html standard. 
The html file consists of six different unique tags that follow the default html standards. As a styler deduce the context above