Create Brain Teaser3
I'm the owner of a school named christland int college. Where we have 20 students, 5 males and 15 females. The males and females are arranged in context of both ordered and unordered forms, where the males and females study three different subjects; math, physics and chemistry. The average score for the male students is 70% in maths, 75% in chemistry and 86% in physics while for the females, the average score in maths is 70%, 72% in physics and 90% in chemistry. Deduce it using the normal HTML principle