Create Brain Teaser4
A malicious user who works in the dark web uses a certain method to trick unsuspecting victims. He started by using 3 different tags by default. The first tags makes reference to the uniformity of linkswhere we have multiple links rangung from 0-9 sent to one email address. The other tags are focused on phone numbers and a text file  is this "For God so loved the world that he gave his one and only Son, that whoever believes in him shall not perish but have eternal life"